Create an abstract class to share trailing whitespace removal logic in the PostProcessor for both regular SQL files and injected SQL.
