+++ src/main/resources/META-INF/plugin.xml
@@ -66,7 +66,7 @@
     <lang.formatter language="DomaSql"
       implementationClass="org.domaframework.doma.intellij.formatter.builder.SqlFormattingModelBuilder"/>
     <preFormatProcessor implementation="org.domaframework.doma.intellij.formatter.processor.SqlFormatPreProcessor" />
-    <postFormatProcessor implementation="org.domaframework.doma.intellij.formatter.processor.SqlPostProcessor" />
+    <postFormatProcessor implementation="org.domaframework.doma.intellij.formatter.processor.SqlFormatPostProcessor" />
     <postFormatProcessor implementation="org.domaframework.doma.intellij.formatter.processor.SqlInjectionPostProcessor" />
 
     <!-- CustomLanguage -->
